Move Vue components to ui
