+++ installer/src/index.js
@@ -1,12 +1,12 @@
 import Vue from 'vue';
-import App from './app.vue';
+import App from './ui/app.vue';
 
 import styles from './styles/index.scss';
 
 const instance = new Vue({
     el: '#app',
     components: { App },
-    template: '<App/>'
+    template: '<App />'
 });
 
 const style = document.createElement('style');
