+++ installer/src/ui/app.vue
