+++ installer/src/ui/destination.vue
