+++ installer/src/ui/intro.vue
