+++ installer/src/ui/license.vue
