+++ installer/src/ui/sidebar.vue
